Fix the Python version select() mechanism to handle PY3-as-default
The target @bazel_tools//tools/python:python_version can be used with select() to determine the Python version in the current configuration. However, it was not updated when --incompatible_py3_is_default was added to switch the default to PY3. This meant that if neither --force_python nor --python_version was passed to the build, but --incompatible_py3_is_default was set, it would incorrectly return PY2 when the actual version was PY3. This CL updates this target's definition to take that flag into account.

Fixes the bug in (but does not close) #7615.

RELNOTES: None
PiperOrigin-RevId: 236642947